* The new error management model is now used everywhere.
* IMPORTANT : there's a very strange bug when using build.sh :
warnings are printed multiple times although the log() method is only
called once on the Logger objects. I'm working on it but if you have
any idea on what may cause that it would be very helpful.

* Despite this nasty bug, since Tom isn't broken (a build.sh all
works) and since this bug doesn't show up in normal uses (I've tested
it "manually" on the example files and the warnings appear only once),
I decided to commit.