A strategy debugger designed to be extended.
How it works : the "decorate" method embbeds every sub-strategy of a strategy
into "Debug" strategies and takes an "observer" as an argument. This observer
should implement two functions, "before" and "after", which are called by the
debug strategies every time they are called, before and after the call of
the strategy they decorate. The debug strategies don't affect current position
so that we can observe things as if they were not present.

We can now imagine a graphical observer ... i'll be working on this.
For now, there is an exemple of a very simple text observer.

Running the example : java ted.StartDebugger