Add a simple test for When_f operator
It uses "visit" for the OldBehaviour test. I think the old "apply" from
MuTraveler should be avoided here, it makes difficult to control failures
as well as understand which code is called.

Also, the new sl "apply" should probably throw VisitFailure or have a way
to report errors, so that we can check both success and failure behavior.
Currently, the sl "apply" behaves like the old apply, and returns indentity
in case of Failure. This is a confusing behavior is "apply" is intended to
replace "visit" in the new world sl