rollback to previous optimizer version because the last one is not sound.
For Radu:

This version is the one you merged so you do not have to bother.
We have written a new version which is much more sound
than this one (the one you've got). However, it led us to discover
some bugs in the PIL code generated by the compilation of non-linear
lists with the current trunk version of the compiler.
The bug occurs because the generated code looks like :

let x = ..
...
letassign x = ..

It appears in lemuridae (in ProofChecker and Proofterm for the
'cutInfo' case) and in the TestListNonLinear unit test.

Since our new optimizer needs a bootstrap to work (because
it uses some very recent gom changes), we would like to
be sure that your new compiler generates sound code before
commiting it.