add jars for antlr 2.6 ant task, and provide them with build.sh. this should fix the build under osx
I don't think this is a good long term solution.
A better solution could be to provide a "toolchain" directory under jtom/
containing ant and the necessary optional plugins, like junit and antlr tasks.
Those cannot be loaded by ant itself, and thus do not belong to stable or src.